pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "ccxt>=1.10.1",
    "cryptography>=45.0.5",
    "openai>=1.93.0",
    "pandas-ta>=0.3.14b0",
    "pandas>=2.3.0",
    "plotly>=6.2.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "socketio>=0.2.1",
    "streamlit>=1.46.1",
    "telegram>=0.0.1",
    "google-generativeai>=0.8.5",
]
